+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,10 +1,12 @@
 object Versions {
     val td = listOf(
-        "1.5.0",
+//        "1.5.0",
         "1.5.1"
     )
     const val tdLatest = "1.5.1"
     const val immutableCollections = "0.3"
+    const val cli = "0.2.0-dev-6"
     const val githubApi = "1.99"
     const val kotlinShell = "0.2"
+    const val dotenv = "5.1.3"
 }