+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/diff/data/TlMetadataDiff.kt
@@ -1,0 +1,7 @@
+package dev.whyoleg.ktd.cli.diff.data
+
+data class TlMetadataDiff(
+    val descriptionList: List<StringDiff>?,
+    val additionList: TlAdditionListDiff?,
+    val propertyList: TlPropertyListDiff?
+)