+++ client/raw/src/main/kotlin/dev/whyoleg/ktd/NativeTarget.kt
@@ -1,5 +1,7 @@
 package dev.whyoleg.ktd
 
+import java.io.*
+
 enum class NativeTarget(val lib: String, val ext: String) {
     Linux("libtdjni", "so"),
     MacOS("libtdjni", "dylib"),
@@ -17,5 +19,20 @@ enum class NativeTarget(val lib: String, val ext: String) {
                 else                     -> error("Target is not supported")
             }
         }
+    }
+}
+
+internal fun link() {
+    runCatching {
+        System.loadLibrary("tdjni")
+    }.recover {
+        val target = NativeTarget.current()
+        val file = File.createTempFile(target.lib, ".${target.ext}")
+        val path = "libs/${target.name.toLowerCase()}/${target.lib}.${target.ext}"
+        TelegramRawClient::class.java.getResourceAsStream(path).copyTo(file.outputStream())
+        System.load(file.absolutePath)
+    }.onFailure {
+        println("Can't load tdlib")
+        throw it
     }
 }