+++ generator/src/main/kotlin/dev/whyoleg/ktd/generator/tl/TlProperty.kt
@@ -1,9 +1,0 @@
-package dev.whyoleg.ktd.generator.tl
-
-data class TlProperty(
-    val name: String,
-    val type: TlType,
-    val descriptions: List<String>,
-    val additions: List<TlAddition>
-)
-