+++ CHANGELOG.md
@@ -1,13 +1,34 @@
 # Changelog
 
+## 0.5.0
+> Unreleased
+- Support Android: armeabi-v7a, arm64-v8a, x86, x86_64
+- Support of tdlib `v1.5.2` - `v1.5.4`
+- Extract `tdlib` in separate dependency `ktd-lib-*`
+  - `ktd-lib-v1.5.0` 
+  - `ktd-lib-v1.5.1`
+  - `...`
+- [TODO] Remove dependencies with latest postfix
+- Rename `TdClient` to `TelegramRawClient`
+- [TODO] Rename `TelegramClient` to `TelegramCoroutinesClient` 
+- [TODO] Split JVM `tdlib` dependency into artifacts per platform
+- Introduce `cli`
+  - Build `tdlib` for supported platform/target
+  - Generate kotlin api
+  - Generate api changelog
+  - [TODO] Configured paths for dependencies and files thorough .env
+- Generate changelog for `v1.5.1` - `v.1.5.4` 
+- Changed dependencies structure
+
+
 ## 0.4.1
-> Published 27 Nov 2019
+> Released November 27, 2019
 - Functions now have nullable `Ref` parameters
 - Fix constructors counting of tl-scheme for `vector`
 - Fix requests cancellation
 
 ## 0.4.0
-> Published 19 Nov 2019
+> Released November 19, 2019
 - Support Linux, MacOS, Win32 and Win64 using JVM
 - Split modules into raw/coroutines client and raw/coroutines api
   - api dependency provide client dependency
@@ -27,7 +48,7 @@
   - `ktd-api-stub` - api stub that can be used in `compileOnly` scope with `ktd-*-client` dependency
 
 ## 0.3.0
-> Published 2 Nov 2019
+> Released November 2, 2019
 - Regenerate api based on tdlib 1.5.1
 - Rework native bridge for more efficient implementation
 - Rework handling of nullability and default values in scheme